requirements.txt
pandas
Flask
numpy
matplotlib
scikit_learn
gunicorn
Werkzeug
Jinja2
itsdangerous
MarkupSafe
azure-devops
